Significance of network intersection

Builds groups of networks randomly and counts their number of intersection edges (intersection size). Each random network is built by randomly sampling from the total number of possible edges. The p-value is computed parameter-free by counting the number of times the intersection size of random networks is smaller than the intersection size in the observed networks and dividing it by the number of iterations. Please see https://github.com/ramellose/anuran for a dedicated tool that assesses the significance of network intersections.

networkIntersectSig(
  rep.num = 5,
  taxon.num = 20,
  avg.network.size = 0,
  inter.size = 0,
  iter = 1000,
  directed = FALSE,
  loops = FALSE,
  hist.rand = FALSE
)

Arguments

rep.num
number of inferred networks

taxon.num
number of taxa in the matrix

avg.network.size
average number of edges in inferred networks

inter.size
number of edges in observed intersection network

iter
number of iterations for random network group construction

directed
if directed, the possible edge number is taxon.num * (taxon.num-1), else it is (taxon.num * (taxon.num-1))/2

loops
if directed, count self-arcs (so compute possible edge number as taxon.num * taxon.num)

hist.rand
if TRUE, plot the histogram of random intersection sizes

Value

p-value

Examples

networkIntersectSig(rep.num=3,taxon.num=8,avg.network.size=12,inter.size=10,directed=TRUE)

#> [1] "Mean random intersections:  0.538"
#> [1] "Standard deviation random intersections:  0.703598643440315"

#> [1] 0.000999001