Cluster taxon- or sample-wise

Cluster sequencing data taxon- or sample-wise. By default, data are clustered sample-wise. The default method is Dirichlet-Multinomial mixtures (DMM) using the DirichletMultinomial package. Note that DMM expects counts and fails if there are taxa that are absent across all samples. Consequently, abundances are scaled to counts and taxa absent across all samples are removed. Also note that counts should not be too large for DMM (not above 10000). If this is the case, they will be scaled (divided by a constant and rounded). In addition, for DMM, the total number of counts should be the same across samples. Partitioning around medoids (PAM) does not expect counts, can deal with absent taxa and large counts. The cluster number can be omitted, in which case the best cluster number is determined in the range provided using a cluster quality index implemented in clusterCrit's function intCriteria.

findClusters(abundances, method = "dmm", k = 3, minK = 2,
  maxK = 10, qualityIndex = "CH")

Arguments

abundances
a matrix with taxa as rows and samples as columns

method
clustering method, supported are dmm and pam

k
cluster number, can be set to NA if cluster number is to be determined with a quality index

minK
the minimum cluster number to test

maxK
the maximum cluster number to test

qualityIndex
CH (Calinski_Harabasz), Dunn, Silhouette and other quality indices supported by clusterCrit's intCriteria function

Value

a cluster membership vector

Examples

data("ibd_taxa")
# unequal sample sums in the example data set are due to the removal of low-prevalence taxa
clusters=findClusters(ibd_taxa,k=NA,maxK=6)

#> Warning: The total counts across samples are not the same. DMM may be biased.

#> Warning: DMM does not accept taxa with no occurrence across samples. These have been removed.

#> Warning: DMM only accepts counts, thus abundances have been scaled to counts.

#> [1] "Abundances have been scaled with factor: 10000"

#> dmn, k=2

#>   Soft kmeans
#>     iteration 10 change 0.001195
#>     iteration 20 change 0.000354
#>     iteration 30 change 0.000147
#>     iteration 40 change 0.000063
#>     iteration 50 change 0.000026
#>     iteration 60 change 0.000011
#>     iteration 70 change 0.000005
#>     iteration 80 change 0.000002
#>   Expectation Maximization setup
#>   Expectation Maximization
#>     iteration 10 change 1.714182
#>   Hessian
#> [1] "DMM goodness of fit:"
#>        NLE     LogDet    Laplace        BIC        AIC 
#> 62140.2542   917.1129 62239.5057 63126.2438 62531.2542 

#> dmn, k=3

#>   Soft kmeans
#>     iteration 10 change 0.021520
#>     iteration 20 change 0.004182
#>     iteration 30 change 0.001771
#>     iteration 40 change 0.001530
#>     iteration 50 change 0.000461
#>     iteration 60 change 0.000086
#>     iteration 70 change 0.000015
#>     iteration 80 change 0.000003
#>   Expectation Maximization setup
#>   Expectation Maximization
#>     iteration 10 change 0.040595
#>   Hessian
#> [1] "DMM goodness of fit:"
#>       NLE    LogDet   Laplace       BIC       AIC 
#> 61977.761  1023.812 61950.250 63458.006 62564.761 

#> dmn, k=4

#>   Soft kmeans
#>     iteration 10 change 0.010407
#>     iteration 20 change 0.002608
#>     iteration 30 change 0.004666
#>     iteration 40 change 0.003470
#>     iteration 50 change 0.001623
#>     iteration 60 change 0.000698
#>     iteration 70 change 0.000338
#>     iteration 80 change 0.000131
#>     iteration 90 change 0.000046
#>     iteration 100 change 0.000016
#>     iteration 110 change 0.000006
#>     iteration 120 change 0.000002
#>   Expectation Maximization setup
#>   Expectation Maximization
#>     iteration 10 change 20.125851
#>     iteration 20 change 11.762338
#>     iteration 30 change 0.000000
#>   Hessian
#> [1] "DMM goodness of fit:"
#>        NLE     LogDet    Laplace        BIC        AIC 
#> 62262.6419   996.8317 62041.5288 64237.1428 63045.6419 

#> dmn, k=5

#>   Soft kmeans
#>     iteration 10 change 0.080885
#>     iteration 20 change 0.003888
#>     iteration 30 change 0.000679
#>     iteration 40 change 0.000096
#>     iteration 50 change 0.000011
#>     iteration 60 change 0.000001
#>   Expectation Maximization setup
#>   Expectation Maximization
#>     iteration 10 change 7.298210
#>     iteration 20 change 0.000000
#>   Hessian
#> [1] "DMM goodness of fit:"
#>       NLE    LogDet   Laplace       BIC       AIC 
#> 62673.551  1136.577 62342.199 65142.308 63652.551 

#> dmn, k=6

#>   Soft kmeans
#>     iteration 10 change 0.006955
#>     iteration 20 change 0.017319
#>     iteration 30 change 0.007230
#>     iteration 40 change 0.000367
#>     iteration 50 change 0.000093
#>     iteration 60 change 0.000023
#>     iteration 70 change 0.000006
#>     iteration 80 change 0.000001
#>   Expectation Maximization setup
#>   Expectation Maximization
#>     iteration 10 change 2.542535
#>     iteration 20 change 0.000003
#>   Hessian
#> [1] "DMM goodness of fit:"
#>       NLE    LogDet   Laplace       BIC       AIC 
#> 62939.437  1110.933 62415.150 65902.449 64114.437 
#> [1] "The optimal cluster number according to Calinski_Harabasz is 2 with value 7.30791430598713"

table(clusters)

#> clusters
#>  1  2 
#> 95 60